Move the code to use types defined based on the targeted ISA
And move some constants to target ISA too

+++ charlotte_core/src/arch/x86_64/memory/page_map/page_table/page_table_entry.rs
@@ -60,7 +60,7 @@ impl PageTableEntry {
         if self.is_present() == false {
             Err(Error::EntryNotPresent)
         } else {
-            Ok(PhysicalAddress::from((self.entry & *ADDR_MASK) as usize))
+            Ok(PhysicalAddress::from(self.entry & *ADDR_MASK))
         }
     }
